+++ src/cli/commands/analyze/dependency-risk-helpers/format-dependency-risks-json.ts
@@ -18,7 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { type DependencyRisksViewModel, PackageIdentity } from './dependency-risks-view-model.ts';
+import {
+  type DependencyRisksViewModel,
+  PackageIdentity,
+} from './view-model/dependency-risks-view-model.ts';
 
 export function formatDependencyRisksJson(project: string, vm: DependencyRisksViewModel): string {
   return JSON.stringify({ project, ...vm }, jsonReplacer, 2);
