+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-table-malware-group.ts
@@ -19,7 +19,7 @@
  */
 
 import { bold, red } from '../../../../../ui/colors.js';
-import type { MalwareGroupVM } from '../dependency-risks-view-model.ts';
+import type { MalwareGroupVM } from '../view-model/dependency-risks-view-model.ts';
 import { genericRiskInfo } from './format-table-risk-generic.ts';
 
 export function appendMalwareGroup(lines: string[], group: MalwareGroupVM): void {
