+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-table-vulnerability-group.ts
@@ -19,12 +19,12 @@
  */
 
 import { bold } from '../../../../../ui/colors.js';
+import type { VersionOptionDescriptionCode } from '../sca-scanner.ts';
 import type {
   FixVersionVM,
   VulnerabilityGroupVM,
   VulnerabilityRiskVM,
-} from '../dependency-risks-view-model.ts';
-import type { VersionOptionDescriptionCode } from '../sca-scanner.ts';
+} from '../view-model/dependency-risks-view-model.ts';
 import { genericRiskInfo } from './format-table-risk-generic.ts';
 
 const MAX_PACKAGE_FIXES = 2;
